Add files via upload
This program allows to simulate 1 dimension movement of similar particles tied up with the same springs. movement starts by withdrawal from a position of equilibrium of one particle. Parameters may be changed to observe different cases. After start program will start the animation of chain of particles and its displacement. It is possible to skip the animation by pressing "esc" if necessary. In the end program show graphics of displacement of some particles from time.